Added file id_list.bin with id's from lists movie and book. created method initializeIdList. Started modifying dialogue-method to make sure 'register' registers id in a format we can search for when the user wants info of certain ID. Getting inputmissmatch while registering Movie working on solution and also adding the same criterias for correct id while registering book.

+++ src/Movie.java
@@ -7,7 +7,7 @@ public class Movie implements Serializable {
 	protected int duration;
 	protected double raiting;
 
-	public Movie(int id, String title, int value, int duration, float raiting) {
+	public Movie(int id, String title, int value, int duration, double raiting) {
 		this.id = id;
 		this.title = title;
 		this.value = value;
